backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.5.0
python-dotenv==1.0.1
numpy==1.26.4
faiss-cpu==1.7.4